Made a huge number of meaningless changes in an attempt to make GCC shut the hell up.  Too many warnings in the build makes us look pretty amateur hour.  Also Fixed a compilation problem in Parse.h.
svn path=/trunk/FreeOrion/; revision=1185

+++ UI/MultiplayerLobbyWnd.cpp
@@ -487,7 +487,8 @@ bool MultiplayerLobbyWnd::PlayerDataAcceptable() const
         empire_names.insert(row.player_data.empire_name);
         empire_colors.insert(row.player_data.empire_color.i);
     }
-    return empire_names.size() == m_players_lb->NumRows() && empire_colors.size() == m_players_lb->NumRows();
+    return static_cast<int>(empire_names.size()) == m_players_lb->NumRows() &&
+	static_cast<int>(empire_colors.size()) == m_players_lb->NumRows();
 }
 
 bool MultiplayerLobbyWnd::CanStart() const
